Added Async and Await functionality
There is a new conditional compile directive called ONLYASYNC that can
be used to only have the Async version of the api available.  This is
useful to find all calls to the mogade api that you want to switch to be
async.  If you still want to be able to use the old api, just don't add
the ONLYASYNC build option.